Time: 0 ms (100%), Space: 8.9 MB (62.83%) - LeetHub

+++ 0900-reordered-power-of-2/0900-reordered-power-of-2.cpp
@@ -3,7 +3,7 @@ public:
     bool reorderedPowerOf2(int n) {
         string s = to_string(n);
         sort(s.begin(), s.end());
-        for(unsigned int i = 1; i < ((unsigned int)1 << 31) - 1; i <<= 1)
+        for(unsigned int i = 1; i < (1ul << 31) - 1; i <<= 1)
         {
             string cmp = to_string(i);
             sort(cmp.begin(), cmp.end());
